app/backend/requirements.txt
Flask==2.3.2
Flask-JWT-Extended==4.5.2
Flask-SQLAlchemy==3.0.3
psycopg2-binary==2.9.6
python-dotenv==1.0.0
gunicorn==20.1.0
pyghmi==1.5.61
nvidia-ml-py3==7.352.0
